requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
shodan==1.30.1
virustotal-api==1.1.11
aiohttp==3.8.5
cryptography==41.0.3
pycryptodome==3.19.0 